# CAN.BackOffice/src/CAN.BackOffice/Agents/BestellingSwagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "Bestellingbeheer Service",
    "description": "Bestellingbeheer Service voor het bestellen van artikelen",
    "termsOfService": "None"
  },
  "basePath": "/",
  "paths": {
    "/api/Bestelling": {
      "put": {
        "tags": [ "Bestelling" ],
        "operationId": "BestellingStatusOpgehaald",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "bestelling",
            "in": "body",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "type": "string" }
          },
          "400": {
            "description": "Client Error",
            "schema": { "$ref": "#/definitions/ErrorMessage" }
          }
        },
        "deprecated": false
      }
    },
    "/api/Sales": {
      "put": {
        "tags": [ "Sales" ],
        "operationId": "BestellingGoedkeuren",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "body",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "type": "string" }
          },
          "400": {
            "description": "Client Error",
            "schema": { "$ref": "#/definitions/ErrorMessage" }
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [ "Sales" ],
        "operationId": "BestellingAfkeuren",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "body",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "type": "string" }
          },
          "400": {
            "description": "Client Error",
            "schema": { "$ref": "#/definitions/ErrorMessage" }
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "ErrorMessage": {
      "type": "object",
      "properties": {
        "foutType": {
          "format": "int32",
          "enum": [ 10, 20, 30, 40 ],
          "type": "integer"
        },
        "foutMelding": { "type": "string" },
        "oplossing": { "type": "string" }
      }
    }
  },
  "securityDefinitions": {}
}